optimize the performance  website, add zod validation to user

+++ src/app/api/category/[id]/route.ts
@@ -4,12 +4,17 @@ import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 
 export async function GET(
-  req: NextRequest,
+  request: NextRequest,
   { params }: { params: Promise<{ id: string }> },
 ) {
   const { id } = await params;
 
   try {
+    const userPayload = verifyToken(request);
+
+    if (!userPayload)
+      return NextResponse.json({ message: "unauthorized" }, { status: 401 });
+
     if (isNaN(Number(id))) {
       return NextResponse.json({ message: "Invalid ID" }, { status: 400 });
     }
@@ -37,13 +42,21 @@ export async function GET(
 }
 
 export async function PUT(
-  req: NextRequest,
+  request: NextRequest,
   { params }: { params: Promise<{ id: string }> },
 ) {
   try {
     const { id } = await params;
 
-    const body = await req.json();
+    const userPayload = verifyToken(request);
+
+    if (!userPayload)
+      return NextResponse.json({ message: "unauthorized" }, { status: 401 });
+
+    if (userPayload.role !== "ADMIN")
+      return NextResponse.json({ message: "access denied" }, { status: 403 });
+
+    const body = await request.json();
     if (isNaN(Number(id)))
       return NextResponse.json({ message: "invalid ID" }, { status: 400 });
 
@@ -53,14 +66,6 @@ export async function PUT(
         { status: 400 },
       );
     }
-
-    const userPayload = verifyToken(req);
-
-    if (!userPayload)
-      return NextResponse.json({ message: "unauthorized" }, { status: 401 });
-
-    if (userPayload.role !== "ADMIN")
-      return NextResponse.json({ message: "access denied" }, { status: 403 });
 
     const updatedCategory = await prisma.category.update({
       where: { id: parseInt(id) },
@@ -95,9 +100,6 @@ export async function DELETE(
   try {
     const { id } = await params;
 
-    if (isNaN(Number(id)))
-      return NextResponse.json({ message: "invalid ID" }, { status: 400 });
-
     const userPayload = verifyToken(req);
 
     if (!userPayload)
@@ -105,6 +107,9 @@ export async function DELETE(
 
     if (userPayload.role !== "ADMIN")
       return NextResponse.json({ message: "access denied" }, { status: 403 });
+
+    if (isNaN(Number(id)))
+      return NextResponse.json({ message: "invalid ID" }, { status: 400 });
 
     await prisma.category.delete({ where: { id: parseInt(id) } });
